urejanje pravic

+++ recipe.php
@@ -72,7 +72,7 @@ $row = $stmt->fetch();
             echo '<div class="vsebina">'.$row['content'].'</div>';
             //preverim ali je trenutno prijavljen uporabnik napisal ta komentar
             if ($_SESSION['user_id'] == $row['user_id']) {
-                echo '<div class="akcija"><a href="comment_delete.php?id='.$row['id'].'" >X</a><div/>';
+                echo '<div class="akcija"><a href="comment_delete.php?id='.$row['id'].'" onclick="return confirm(\'Prepricani?\')" >X</a><div/>';
             }
 
             echo '</div>';
